+++ client/src/components/lost/LostPostMap.js
@@ -1,23 +1,9 @@
 import React, { useEffect, useState, useRef } from 'react';
-import {
-  Map,
-  MapMarker,
-  Circle,
-  Polyline,
-  CustomOverlayMap,
-} from 'react-kakao-maps-sdk';
+import { Map, MapMarker } from 'react-kakao-maps-sdk';
 import PropTypes from 'prop-types';
 import FindPlaceName from './FindPlaceName';
+import DrawCircle from './DrawCircle';
 
-function DistanceInfo({ distance }) {
-  const totalRadius =
-    distance >= 1000 ? `${(distance / 1000).toFixed(1)}km` : `${distance}m`;
-  return (
-    <ul className="relative bg-white top-1 left-1 p-2 m-0 text-xs border-1 border-solid shadow rounded">
-      <li>총반경 {totalRadius}</li>
-    </ul>
-  );
-}
 function FindLocation({ setAddress, setAddressName, setRadius }) {
   const [position, setPosition] = useState({
     center: {
@@ -79,6 +65,7 @@ function FindLocation({ setAddress, setAddressName, setRadius }) {
     setAddress(position.center);
     alert('위치 등록이 완료되었습니다. 지도를 닫아주세요.');
   };
+
   useEffect(() => {
     if (navigator.geolocation) {
       // GeoLocation을 이용해서 접속 위치를 얻어옵니다
@@ -121,63 +108,18 @@ function FindLocation({ setAddress, setAddressName, setRadius }) {
         onMouseMove={handleMouseMove}
       >
         {isDrawing && (
-          <>
-            <Circle
-              center={position.center}
-              radius={position.radius}
-              strokeWeight={1} // 선의 두께입니다
-              strokeColor="#00a0e9" // 선의 색깔입니다
-              strokeOpacity={0.1} // 선의 불투명도입니다 0에서 1 사이값이며 0에 가까울수록 투명합니다
-              strokeStyle="solid" // 선의 스타일입니다
-              fillColor="#00a0e9" // 채우기 색깔입니다
-              fillOpacity={0.2} // 채우기 불투명도입니다
-            />
-            <Polyline
-              path={[position.center, mousePosition]}
-              ref={drawingLineRef}
-              strokeWeight={3} // 선의 두께 입니다
-              strokeColor="#00a0e9" // 선의 색깔입니다
-              strokeOpacity={1} // 선의 불투명도입니다 0에서 1 사이값이며 0에 가까울수록 투명합니다
-              strokeStyle="solid" // 선의 스타일입니다
-            />
-            <CustomOverlayMap
-              position={mousePosition}
-              xAnchor={0}
-              yAnchor={0}
-              zIndex={1}
-            >
-              <DistanceInfo distance={Math.floor(position.radius)} />
-            </CustomOverlayMap>
-          </>
+          <DrawCircle
+            position={position}
+            mousePosition={mousePosition}
+            drawingLineRef={drawingLineRef}
+          />
         )}
         {circle && (
-          <>
-            <Circle
-              center={circle.center}
-              radius={circle.radius}
-              strokeWeight={1} // 선의 두께입니다
-              strokeColor="#00a0e9" // 선의 색깔입니다
-              strokeOpacity={0.1} // 선의 불투명도입니다 0에서 1 사이값이며 0에 가까울수록 투명합니다
-              strokeStyle="solid" // 선의 스타일입니다
-              fillColor="#00a0e9" // 채우기 색깔입니다
-              fillOpacity={0.2} // 채우기 불투명도입니다
-            />
-            <Polyline
-              path={[circle.center, circle.mousePosition]}
-              strokeWeight={3} // 선의 두께 입니다
-              strokeColor="#00a0e9" // 선의 색깔입니다
-              strokeOpacity={1} // 선의 불투명도입니다 0에서 1 사이값이며 0에 가까울수록 투명합니다
-              strokeStyle="solid" // 선의 스타일입니다
-            />
-            <CustomOverlayMap
-              position={circle.mousePosition}
-              xAnchor={0}
-              yAnchor={0}
-              zIndex={1}
-            >
-              <DistanceInfo distance={Math.floor(circle.radius)} />
-            </CustomOverlayMap>
-          </>
+          <DrawCircle
+            position={circle}
+            mousePosition={circle.mousePosition}
+            drawingLineRef={drawingLineRef}
+          />
         )}
 
         {position.center && (
@@ -227,7 +169,4 @@ FindLocation.propTypes = {
   setRadius: PropTypes.number.isRequired,
 };
 
-DistanceInfo.propTypes = {
-  distance: PropTypes.number.isRequired,
-};
 export default FindLocation;
